Splitted the server cast service into partial classes
